Authenticate clients using SASL PLAIN.
This is a first version of SASL based authentication using SASL. For
now, all passwords are accepted as we are only interested in determining
the client identity as a result of the SASL handshake instead of simply
using the AMQP client container name. Next step should be to verify
credentials against an (external) identity provider.