Replace EventBus based message passing with direct method invocation.
The TelemetryEndpoint has forwarded downstream messages to the
TelemetryAdapter via the Vert.x EventBus and a SharedMap. This has
required a lot of JSON message creation/parsing/passing but did not
really have any big advantage regarding decoupling of the two
components.

The TelemetryEndpoint has therefore been refactored to directly invoke
methods on a DownstreamAdapter to pass on client life-cycle events and
downstream messages. The overall complexity of forwarding a message to a
downstream container has been greatly reduced by this step. In addition,
there are no messages being "buffered" in the shared map anymore which
makes the HonoServer more resilient against failing connections.