Put all default/example configuration into single Docker image.
The configuration files for the Dispatch Router and Hono server are now
copied into a single hono-default-config Docker image in order to
simplify start-up of the Hono example instance.

The configuration of the Dispatch Router has been merged into a single
JSON file (qdrouterd.json) which now also contains the vhost
configuration inline. This better supports the usage of Docker secrets
(introduced with Docker 1.12) for providing a configuration file on the
Dispatch Router container's command line.