Prevent upstream client from starving.
When the downstream container does not replenish the downstream sender
with credits for some time, the upstream client may run out of credits
even if the downstream sender has remaining credit to forward messages
downstream.
The upstream client is now potentially replenished with credit on every
message processed if the downstream sender has credit remaining.