Refactor AuthorizationService to define hasPermission method only.
The authorization service's purpose is to support authorization
decisions that need to be made in Hono's arbitrary components. The
management of authorization rules is not part of this responsibility. In
fact, it should be easy to provide an alternative implementation of the
authorization service, e.g. by creating a facade for existing security
infrastructure without the need to implement management operations
(which will never be invoked anyway in this case).

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>