[#182] Close connection with client on token expiration.
Proton-vertx currently does not support authentication using a token
which we could use to refresh an existing token with the authentication
server. For the time being we therefore close the connection with the
client when the token expires, thus triggering the client to re-connect
and authenticate again, resulting in a new token being issued.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>