[#231] Remove sender creation locks when connection fails.
The locks indicating an ongoing attempt to open a sender link are now
removed if the connection to the server fails during such an attempt.

This way, the next attempt to open a sender (once the connection has
been re-established) does not fail due to an attempt already being in
progress.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>