Use ApplicationRunner.run() for deploying service Verticles.
The Spring Boot ApplicationRunner.run() method is invoked when Spring
Boot's start up of the application is finished. It is therefore a better
place to deploy the service Verticles than using an individual Spring
beans's @PostConstruct annotated method which may be invoked at a time
when not all beans have been initialized properly.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>