Allow to re-use the ApplicationConfig with other config types
This change does refactor the device registry module in order to
allow re-using the same ApplicationConfig class with different device
registry implementations. Right now the method
'registrationAssertionFactory' limits this so "File*" implementations.

With the change applied the configuration is no longer actively
looked up by the method, but injected by the IoC framework and thus
can be swapped. Also is only the "signature" part required and so the
interface 'SignatureSupporting' was added.

Signed-off-by: Jens Reimann <jreimann@redhat.com>