Make readiness check less stressful for Authentication service.
The readiness checks of Hono Messaging and Device Registry have been
establishing a connection to the Authentication service as part of their
check. This put unnecessary stress on the Authentication service and has
been replaced with a simple DNS lookup on the Authentication service
host which serves the same purpose of making sure that an Authentication
service instance is in "ready" state.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>