[#426] Add support for PKCS#1
This changes enables the use of PKCS#1 files using RSA keys. This is an
extension to the current PKCS#8 logic in the KeyLoader. Vertx and the
network functionality of vertx already support PKCS#1 through the
PEM options.

In order to make this work this changes also adds a PEM parser and a
set of unit tests to test the new functionality and a few corner cases.
It simplifies the logic of detecting file types and also allows to
explicitly specify the file format through the configuration for cases
when the file type cannot be detected due to an unknown extension.

Signed-off-by: Jens Reimann <jreimann@redhat.com>