Properly handle a failed initialization
This change does re-write the startup code in way that it does handle
a failed initialization chain. It removed an un-necessary compose
element at the end of the chain and uses a CompletableFuture instead
of the CountDownLatch to wait for the final result.

This way the final result may receive any exception thrown by the
startup chain of futures and immediately shutdown outside of the vertx
context.

Signed-off-by: Jens Reimann <jreimann@redhat.com>