[#443, #442] Move disposition handling to MessageSender implementation.
So far some of the MessageSender interface's methods allowed to pass in
a disposition handler which was responsible for handling the delivery
updates from a downstream container. This way, the delivery semantics of
the particular sender implementation (Telemetry vs. Event) had to be
provided/implemented clients using MessageSender instances. From an
architectural point of view this didn't make much sense because it
clearly contradicted the principle of encapsulation.

In order to move the responsibility for implementing the delivery
semantics back to the MessageSender implementations, the MessageSender's
operations have been changed to no longer support the passing in of a
disposition handler. In addition, all methods now return a Future
indicating the outcome of the operation according to the
implementation's specific delivery semantics, i.e. the
TelemetrySenderImpl now returns a succeeded future if credit is
available without waiting for the message being settled and accepted
whereas the EventSenderImpl now returns an incomplete future which will
complete successfully once the message has been accepted by a downstream
peer.

This way, clients using MessageSender no longer need to bother about the
nitty-gritty of how to implement the delivery semantics correctly but we
have instead encapsulated the disposition handling in one central place
(where it belongs).