[#457] Clearing out connection after loss of connection
This change sets the connection to 'null' when it is being handled is
disconnected. This will ensure future checks on the connection won't
consider it usable.

There is also a change in the
DisconnectHandlerProvidingConnectionFactory which takes care of first
sending out the events and then updating the latch. Which might cause a
race condition issue between the unit test and the connection factory.
If the latch is updated in the connection factory, the handler still
is not executed and the client will still be disconnected. So if the
following call to "getOr" is "too quick" then it will fail.

Signed-off-by: Jens Reimann <jreimann@redhat.com>