Do not support command chaining in HonoClient anymore.
Most methods of HonoClient had returned the client instance so far in
order to support command chaining.
However, there was nothing much to chain because the effect of invoking
the (asynchronous) operations is usually not immediately visible in the
client, e.g. when invoking the "connect" method, the returned client
will most likely not (yet) be connected. Returning void instead thus
prevents users from (incorrectly) assuming that command chaining works
in the usual way.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>