[#465] Condensed user guide page weights.
Previously there was not enough room for separating the device registry
setup page from the device registry RESTful resources pages.
The gaps are now bigger for the protocol adapters and the core components
for future pages to be added.
The RESTful device registry resources page is now inside the defined
range.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>