[#388] Removed null check for tenantId in reply helper method.
The check for null is not appropriate for the reply to a incorrect request
that does not contain a tenantId. The best way to reply is to set the
tenantId in the reply to what it was in the request: null.
So the check for null was removed (again) and instead the documentation
of the param is extended.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>