[#507] Forward downstream disposition to upstream client.
When forwarding unsettled telemetry messages to the downstream AMQP
Messaging Network, the adapter now forwards the disposition received
from the downstream container to the upstream client.

For pre-settled messages the behavior is unchanged, i.e. the adapter
immediately accepts and settles the message with the upstream client
before forwarding the message downstream.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>