[#518] Fix permission check for Credentials API actions.
The introduction of an enum type for defining request response actions
implictly changed the name of the message subjects that were checked
with the permissions defintion. The previous name "get" was changed to
"ACTION_GET" and was not matching with the permissions definition (which
sticks to use "get"). The enum members are renamed and now match again
with the permissions definition.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>