[#524] Introduce response cache for Tenant API.
Protocol adapters need to access the tenant information via the Tenant
API quite frequently. To minimize the number of requests that are sent
to the Tenant service (e.g. inside the device registry), the results are cached per tenant
(cache-key: tenantId). Their default timeout is 10 minutes, which can be
configured to a different value.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>