Include correlation info in event bus messages.
The AMQP based endpoints so far have received response messages frmo
service implementations as direct replies to the requests sent via the
event bus. The reply has then been enriched with correlation information
from the original AMQP request message and been sent to the reply-to
address via the event bus.

We now include all information required for correlating responses with
requests in the request message sent to the service implementation. The
services then create response messages which also include this
correlation information and send the response directly to the reply-to
address via the event bus, thus saving the additional hop.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>