[#558] Don't mount Kubernetes secrets to /run/secrets.
With the latest kubernetes version it is no longer possible to mount
secrets to the /run/secrets path. The configuration yaml files are now
copied to the target/deploy folder using the Maven resources plugin
using deployment target specific properties for Docker Swarm and
Kubernetes. For kubernetes the secrets are now mounted to /etc/hono
while the secrets on Docker Swarm are still mounted to /run/secrets.

Fixes #558.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>