[#586] Cleanup how instances of HonoClientImpl are created
This change does clean up the way how new instances of HonoClientImpl
can be created. This introduces a new method on the HonoClient instance
which will create a new instance of a default implementation of
HonoClient based on HonoClientImpl.

The reason for this changes from from #586 which is based on the problem
the internal "impl" package was not exported in OSGi and so directly
accessing the HonoClientImpl constructor was not possible.