[#349] Do not fail TLS handshake if tenant is disabled.
When a device authenticates using a client certificate then the tenant
that the device belongs to might be disabled. We now let the TLS
handshake succeed in this case and let the protocol adapter decide
whether it wants to reject the device's connection attempt or not. This
allows for a more consistent interface towards the device which will
receive an error at the transport protocol layer in any case.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>