[#605] Do not fail immediately on SASL handshake failure.
HonoClient now tries to re-connect as many times as indicated in the
ProtonClientOptions regardless of the reason for connection failure
before giving up. If the last failure is due to a failed SASL handshake
then the connection result handler will be failed with a
ClientErrorException and status 403, otherwise it will fail with a 503
ServerErrorException.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>