[#319] Add high level tracing to HTTP and MQTT adapters.
All standard adapters are instrumented using OpenTracing in order to be
able to track the processing of messages that have been uploaded by
devices.

In this first step, only the handlers for incoming HTTP requests and
published MQTT messages are instrumented, i.e. we only trace the
incoming request and the outcome of processing the request at a very
high level.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>