[#649] Improve connection attempt behavior.
The client's "connect" methods by default will only try three times to
establish a TCP connection with the peer before giving up. Based on the
value of the new "reconnectAttempts" property in ClientConfigProperties,
the client will then either re-try to connect to the peer (including a
fresh lookup of the peer's host name) or fail the overall connection
attempt. This way, the client will not get stuck in an endless loop if
the peer's IP address has changed or the peer has crashed while the
client tries to connect to it.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>