Require downstream links to be opened during connection establishment.
When a device connects and is authenticated, the adapter will
providently try to open all relevant downstream links (Registration,
Telemetry and Event APIs). The connection will only be established, if
the links can be established successfully. This way, the links will be
available (and most likely have received some credit) when the device
publishes its first message.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>