Improve error handling of HTTP adapter and HTTP based endpoints.
A DefaultFailureHandler is now registered on the service's Router to
handle ServiceInvocationExceptions that have caused a RoutingContext to
fail. The handler creates an HTTP response corresponding to the
exception's fields.

The HttpUtils methods now fail the routing context with a
ServiceInvocationException instead of just a status code. This way we
can convey additional information to the client, e.g. the detail message
that is contained in the exception.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>