Prevent race condition when opening MQTT connection with devices.
When the MQTT adapter sent out its CONNACK packet to a device, it was
sent on the vert.x context of one of the HonoClient instances used for
authenticating the device. The MqttServer sets a boolean flag on the
MqttEndpoint to mark it as connected after it has sent out the packet.

When the device publishes its first message, the PUBLISH packet is
received on the MQTT adapter's context (which is usually running on a
different thread than the HonoClient context). In rare cases this has
caused a race condition when the flag's updated value ("connected") was
not yet visible on the current thread and the connection to the device
has therefore been closed by the MqttServer.

In the MQTT adapter we now make sure that the CONNACK packet is sent on
the same adapter's context so that the race condition cannot occur
anymore.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>