[#696] Use current vert.x Context for connecting to peer.
HonoClientImpl now uses the current vert.x Context that the client is
running on for connecting to and interacting with the peer. This way
most of the invocations on HonoClient can be executed immediately
instead of scheduling a task on the HonoClient's Context. We also make
sure that the Futures returned by HonoClient operations are
completed/failed on the client's Context, i.e. we are no longer
switching Context when executing handlers registered on the Future.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>