[#630] Cancel command expiry timer in HTTP adapter.
If a command was received or the HTTP connection was closed, a created
expiry timer is now cancelled explicitly. Before the timer stayed alive until
it fired - which could be a problem with many very long running timers.
A handler that called the receiver link close a second time was removed.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>