First integration test support for Command and Control.
This commit tests the recognition of the ttd parameter for the HTTP
adapter with telemetry messages and verifies that the ttd is contained in
the received message.
It then sends a command in the http response.
Not contained is the response to the command - this will follow in next
commits.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>