Do not share CommandResponseSender instances.
Applications may use the same reply-to address when sending commands to
multiple devices. If protocol adapters share a sender link for the same
target address and close the link when they have sent the message, then
the link might be rendered unusable for another device's response that
needs to be sent to the same reply-to address.

For the time being, the CommandResponseSender links are therefore no
longer shared in order to prevent this from happening. In the future we
might come up with a decent way of safely sharing the links again, e.g.
using a reference counter to determine if a link can actually be closed.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>