Free link resources for HonoClient if remote peer did not send detach.
To prevent a link based memory leak, sender and receiver links are
currently freed in their close handlers. However, the close handler
is only called if the remote peer reacts with a detach frame after the
client closed the link. Now a timer is created additionally to
supervise if there should be no detach frame from the remote to free
the link resources then.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>