[#794]: Replace Dropwizard with Micrometer
This change replaces the internal usage of Dropwizard in favor of
Micrometer. This also makes use of using tags for meters.

Micrometer is integrated using Spring Boot, and vertx is only configured
to contribute to that existing meter registry. This allows to take full
advantage of the Spring Boot Micrometer integration, and simply drop in
any Micrometer adapter, including the use of auto-configuration.

With this change Hono can be operated in two modes, either the "legacy"
mode, which provides at least the same metrics as before, publishing
over the same Graphite reporter, with exactly the same metrics names.

Or using the new format, which is the default and makes use of any
Micrometer backend implementation that is dropped into the classpath.