[#819] Fix permanent blocking of creation locks for Hono clients.
In very rare situations there was the chance to permanently block the
creation of Hono clients that did not recover anymore. To improve the
behaviour of parallel client creation, the boolean locks are now based on
timestamps that will expire after one second, so a new client can be
created.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>