Hono example now handles multiple ttd events at startup correctly.
When connecting devices by a connection oriented protocol with ttd values
of -1 or 0, the handling when starting the application after the device
was connected and disconnected for an arbitrary number of times was
incorrect. This commit handles the notifications coming in with a delay
of one second and only considers the last event that came in during this
delay interval.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>