Simplified example app and always get a fresh command client.
Previously, for notifications with ttd -1 a periodic timer was started
inside the retrieval of a command client, and was directly canceled
and restarted again before the next command was sent.
The timer now is correctly placed outside of the retrieval code of the
command client, which corrected the behaviour and additionally simplifies
the code.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>