Do not return cached CommandConsumer instances.
The getOrCreatedCommandConsumer method has been renamed to
createCommandConsumer and will no longer return an already existing
instance of a command consumer for a given device but will instead fail
the returned future with an
org.eclipse.hono.client.ResourceConflictException to indicate that a
consumer for the given device is already in use.

The original behavior allowed an implementation to return a consumer
that was scoped to another message handler than the one passed into the
method as an argument. However, client code had no chance to determine
whether it got back a newly created instance or an existing one. This
has been resolved with the new method semantics.

There is now also an integration test for the HTTP adapter which
verifies the correct behavior when a device issues multiple requests
with a TTD without waiting for a response before sending the next
request.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>