[#669] Provide one-way pattern for commands.
The command client has a new method for sending a command:
sendOneWayCommand with 2 different signatures.
Internally this is controlled by a specific value in the application
properties that results in a message lacking the reply-to property.
Adapters (not contained in this commit) should complete the command
successfully after having received and approved it without waiting for
the response of a device.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>