[#830] Re-create AMQP adapter's command consumers on connection loss.
The AMQP adapter now creates command consumer links which will be
checked periodically for liveness and which will be automatically
re-created if the command connection is lost.

For now, the adapter uses a default check interval of 10 seconds.
This should probably be made configurable in the future or, even better,
should be determined based on the device's AMQP connection idle-timeout.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>