Log the correlation ID of a command message.
The correlation ID is the message-id or correlation-id from the AMQP
message that has been used to convey the command. Logging the (derived)
command request ID doesn't make sense in the CommandConsumer because it
is only used by specific protocol adapters (which may log it once they
are actually using it).

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>