Make disposition handling a terminal operation in CommandContext.
The command message received from an application needs to be settled by
a protocol adapter once it has processed the message. The corresponding
methods of CommandContext now also accept an integer indicating the
number of credits to flow to the sender and also finish the OpenTracing
span contained in the context.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>