Rename example consumer application to `HonoExampleApplication`.
Earlier this has been a purely consumer based example application. Since we
introduced Command and Control, it supports more and more sending commands
as a reply to a received downstream message, so the original name did
not match anymore and already led to confusion in the community.
Thus it is renamed to a hopefully better name now.

Signed-off-by: Karsten Frank <Karsten.Frank@bosch-si.com>