Fix HTTP adapter's handling of TTD.
The HTTP adapter now includes the TTD provided by a device in the
downstream message only if it can successfully acquire a Command
consumer for the device. If the consumer for the device is already in
use and the request contains an empty notification, no message will be
sent downstream.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>