[#130] Support configuration of max message size in AMQP adapter.
Also make max frame size and session window size configurable.
The max message size is still not enforced by the adapter. However, the
max message size is conveyed to the device in the adapter's attach frame
so that the device knows the limit. On the server side, the maximum
memory consumption of unsettled messages can be limited by means of the
configured max session window size.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>