[#931] Check or set connection limit in MQTT adapter
If no connection limit is configured, the MQTT protocol adapter auto-
configures a limit based on the amount of memory.
If a limit is configured and it is above the recommended value, a warning is printed.

Renamed the property `maxConcurrentConnections` to `maxConnections`.
It is now initialized with 0, which communicates that no limit has been configured.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>