[#732] Add example configuration for client certs + create new certs.
The script for creating the demo certs creates now a client certificate
for device 4711 as well. The example tenant configuration contains now
the public key of the CA cert as trust anchor for DEFAULT_TENANT.
The script for creating the demo certs creates writes the CA's public key
into a properties file. The build process of the module `deploy` sets
the subject and the public key from the properties file as the trust anchor of
DEFAULT_TENANT in the file `example-tenants.json`.
Similar the subject of the client certificate of device 4711 is set in
the example credentials file.
Created new certificates + add AMQP adapter to script create_certs.sh
+ remove certificates for 'Hono Messaging' + update README.md of module
`demo-certs`.
Added examples to the User Guides of HTTP adapter and MQTT adapter.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>