[#931] Add `ConnectionLimitManager` + Metrics.getNumberOfConnections() returns int.
Add the `ConnectionLimitManager` and use it for MQTT and AMQP protocol
adapters.
AMQP: If no connection limit is configured, the AMQP protocol adapter auto-
configures a limit based on the amount of memory and the configured
maximum session window size. If a limit is configured and it is above
the recommended value, a warning is printed.
The MQTT adapter uses now ConnectionLimitManager.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>