[#1083] Unified file paths in documentation, removed shell prompts.
Prompts are completely removed in front of shell commands to enable easy
copy & paste. Paths are usually relative to the base directory of the
Hono repository. Where relevant the directory in which a command needs
to be executed is given in a comment.
Some commands (starting adapters with Spring Boot) are fixed.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>