[#780] Implement DownstreamSenderFactoryImpl.
The DownstreamSenderFactory methods have been moved from
HonoConnectionImpl to DownstreamSenderFactoryImpl.
As part of this, the org.eclipse.hono.client.impl.*ClientImpl hierarchy
has been refactored to use a HonoConnection instance instead of passing
around multiple properties of the connection explicitly. This serves as
the groundwork for implementing the other client factories analogously
to the DownstreamSenderFactory.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>