[#930] Add tenant-scoped command consumer.
This introduces a new command consumer on a tenant-scoped address.
This consumer handler will map the device id in the command message
'To' address to a corresponding 'via' device id (if applicable) and
delegate the command either to a local command handler or to a consumer
on another protocol adapter instance by way of the AMQP network.
As the existing command consumer address scheme isn't changed (yet),
this change doesn't cause any incompatibilities.

Signed-off-by: Carsten Lohmann <carsten.lohmann@bosch-si.com>