Change the model of the credentials.
This change alters the model of the device registry HTTP API in way
that redundant parts of the secrets (type, auth-id, enabled) get pulled
up one level. And that new level gets the "ext" extensions as well.

The "enabled" flag is still present on the "secrets" level, as it might
be beneficial to e.g. disable a single secret temporarily.