Improve the selection of documentation versions.
Versions of the documentation are now mapped on languages for Hugo,
allowing to leverage the multilingual features of Hugo. The current
stable version is a level above the other versions, improving the
search engine ranking of this version. Version "stable" has the URL
path "/hono/docs/", nightly master has "/hono/docs/dev/" and e.g.
version 1.2.3 would be under "/hono/docs/1.2/". Having a version marked
as stable is required for this to work.
The version is now selected in the documentation itself and no longer
on the homepage.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>