[#1414] Use standard Prometheus/Grafana charts instead of operator.
The chart no longer relies on the Prometheus Operator. Instead, the
standard Prometheus chart is used to deploy a Prometheus instance if
enabled. Alernatively, the chart supports using an existing Prometheus
server by means of specifying a host name and port.

The chart now uses the standard Grafana chart to create a Grafana server
instance if enabled.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>