[#1188] Close inactive downstream sender links.
A timeout can now be configured for inactive AMQP links. If no message
is sent for the configured time period, the link is closed. Each time
the a message is sent on this link the timestamp in stored in the
attachments of the link. A timer checks regularly that this timestamp
is not older than the configured timeout. If no timeout is configured
the link will not be closed automatically.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>