[#1227] Client publishes link close events on the Vert.x event bus.
The Hono client now publishes an event on the event bus in the
`closeLinks` method. This decouples the usage of the client from it's
inner logic. It allows the applications to get informed of link close
events without the need to propagate a handler through multiple layers
and allows for an arbitrary amount of subscriber to this events.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>