[#1227] Remove metrics (telemetry and connections) for inactive tenant.
The interface `Metrics` has a new method to remove metrics for telemetry
(and event) messages.
The `AbstractProtocolAdapterBase` registers a listener for the sender
link  close event at the event bus and removes the telemetry metrics for
the corresponding tenant.
The class `MicrometerBasedMetrics` automatically removes the metrics of
(authenticated) connections for a tenant when the count drops to zero.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>