[#1188] Close AMQP links when tenant timeout is exceeded.
A timeout for tenants can be configured (tenantIdleTimeout). When there
has been no communication between a protocol adapter instance and the
devices of a tenant, the adapter releases closes the AMQP sender links
of that tenant.

The class MicrometerBasedMetrics monitors connects and disconnects
as well as massage sending for each of the tenants it has seen and
publishes a message on the event bus when the timeout occurs.
The (caching) client factories register at the event bus for tenant
timeout events. When they receive the timeout event they close the
links of the client for this tenant and remove the client from the
cache.
If no timeout is configured the link will not be closed automatically.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>