Improve Getting started Guide part that covers C & C.
The second part of the Getting started Guide, which explains Command and
Control, didn't work anymore because he expected Hono on localhost.
There were also some redundancies because first the sample application
was used and then the same thing was done again with the CLI.
Now only the CLI is used and the explanation uses the MQTT adapter.
In addition, the log output of the CLI has been improved a little.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>