[#623] Use Tenant instead of TenantObject as in-memory data model.
The example device registry now uses Tenant instances for keeping tenant
data in memory. This is more aligned with the fact that the Tenant
object is used as the data model for the Tenant related operations of
the Device Registry Management API.
It also makes it easier to project on specific properties of the Tenant
in the result set of Tenant API operations (which are based on
TenantObject as the data model).

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>