Disable swapping for containers in integration tests.
It seems that the memory limit configured for the Device Registry is not
always sufficient for running the integration tests. This is usually not
a problem when running locally, with swap space available, but seems to
have led to spurious failures on Travis, where the adapters suddenly
loos connection with the Device Registry because it crashed (due to OOM
IMHO). Swapping has therefore been disabled for the containers
altogether and the Device Registry's memory limit has been slightly
increased.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>